requirements.txt
pyaes
tinytuya
mss
opencv-python
numpy
python-dotenv
